+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+
 ### Added
 - Raise a RuntimeError if there are more than 5000 levels in a column (#42)
 - Emit a warning if the column levels during transform don't overlap