+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+
 ### Added
+- Raise a RuntimeError if there are more than 5000 levels in a column (#42)
 - Emit a warning if the column levels during transform don't overlap
   at all with levels from fitting (#41)